refactor(stella-cli): split command_deck.rs's slash/turn/support helpers into sibling modules
command_deck.rs sat at exactly its 3737-line grandfathered ceiling — zero
headroom for the next change. It had already established the pattern for
this (20 existing sibling modules); this continues it for the remaining
free functions that were left inline: the slash-command dispatcher and
installed-agents ops move to slash_commands.rs, the lead turn driver to
lead_turn.rs, the two SETTINGS-panel snapshot builders to panel_snapshots.rs,
the channel-owning background tasks (MCP connect, session replay, PR/inbox
pollers, supervisor-message handling) to driver_support.rs, and the
ISSUES-tab entity-hit assemblers into issues.rs beside their only caller.

No logic changed — every move is a relocation plus visibility annotations.
command_deck.rs drops from 3737 to 2474 lines (34%), with 1200+ lines of
real headroom now instead of none.
