+++ crates/stella-cli/src/command_deck/tests.rs
@@ -1,7 +1,7 @@
-// Imported here rather than inherited through `use super::*`: the label
-// belongs to the deck's ask io, which lives in `mid_turn_ask` now, and this
-// module is its only remaining user in `command_deck`'s namespace.
-use crate::interactive::FREE_TEXT_LABEL;
+// Imported here rather than inherited through `use super::*`: the label and
+// the trait belong to the deck's ask io, which lives in `mid_turn_ask` now,
+// and this module is their only remaining user in `command_deck`'s namespace.
+use crate::interactive::{AskUserIo, FREE_TEXT_LABEL};
 
 use super::pr_observe::scrub_gh_command;
 use super::skills::{
@@ -30,35 +30,6 @@ async fn pr_observer_preserves_github_auth_only() {
         String::from_utf8_lossy(&output.stdout),
         "unset|repository-secret|unset|visible"
     );
-}
-
-#[test]
-fn deck_arg_commands_parse_models_forms_and_leave_sentences_as_prompts() {
-    assert!(matches!(
-        parse_models_command("/models refresh"),
-        Some(ModelsCommand::Refresh { force: false })
-    ));
-    assert!(matches!(
-        parse_models_command("/models refresh --force"),
-        Some(ModelsCommand::Refresh { force: true })
-    ));
-    assert!(matches!(
-        parse_models_command("/models list"),
-        Some(ModelsCommand::List)
-    ));
-    // One unrecognized token is a typo'd subcommand → usage, never a
-    // model call; a sentence stays a prompt.
-    assert!(matches!(
-        parse_models_command("/models refrsh"),
-        Some(ModelsCommand::Usage(_))
-    ));
-    assert!(parse_models_command("/models what can I use").is_none());
-    // Bare forms and non-command paths are not arg commands — and the
-    // removed `/model-<role>` heads no longer parse (model config lives
-    // on the SETTINGS tab).
-    assert!(parse_models_command("/models").is_none());
-    assert!(parse_models_command("/model-default zai/glm-5.2").is_none());
-    assert!(parse_models_command("/src/main.rs explain").is_none());
 }
 
 #[test]
@@ -472,142 +443,6 @@ fn stray_stop_and_hold_is_a_no_op() {
 }
 
 // ISSUES tab: entity-hit assembly
-
-#[test]
-fn agent_entity_hits_filter_by_name_or_description_case_insensitively() {
-    let entries = vec![
-        stella_tui::InstalledAgentEntry {
-            name: "reviewer".into(),
-            description: "Reviews diffs".into(),
-            tools: None,
-            scope: AgentScope::Project,
-            source_path: String::new(),
-            version: 1,
-            versions: vec![],
-            content: String::new(),
-        },
-        stella_tui::InstalledAgentEntry {
-            name: "planner".into(),
-            description: "Plans work".into(),
-            tools: None,
-            scope: AgentScope::User,
-            source_path: String::new(),
-            version: 1,
-            versions: vec![],
-            content: String::new(),
-        },
-    ];
-    let hits = agent_entity_hits(&entries, "REVIEW");
-    assert_eq!(hits.len(), 1);
-    assert_eq!(hits[0].kind, "Agent");
-    assert_eq!(hits[0].insert, "reviewer");
-    // Description text matches too; the empty query matches all.
-    assert_eq!(agent_entity_hits(&entries, "plans")[0].label, "planner");
-    assert_eq!(agent_entity_hits(&entries, "").len(), 2);
-}
-
-#[test]
-fn memory_hits_carry_the_preview_provenance_and_citation_suffixes() {
-    let hit = memory_hit(
-        "naming-convention",
-        "Prefer kebab-case for  skill names\nand slugs.",
-        "2026-07-01T00:00:00Z",
-        Some((12, 0.9)),
-    );
-    assert_eq!(hit.kind, "Memory");
-    assert_eq!(hit.insert, "naming-convention");
-    assert_eq!(
-        hit.description,
-        "Prefer kebab-case for skill names and slugs. · observed \
-         2026-07-01T00:00:00Z · cited 12× avg 0.9"
-    );
-    // Observation time is the only time a node carries: no `valid from`
-    // clause restating it (#3136).
-    assert!(!hit.description.contains("valid from"));
-
-    // No citations → no suffix; a long content truncates char-safe with an
-    // ellipsis.
-    let long = "x".repeat(200);
-    let hit = memory_hit("m", &long, "2026-07-01", None);
-    assert!(
-        hit.description
-            .starts_with(&"x".repeat(MEMORY_PREVIEW_CHARS - 1))
-    );
-    assert!(hit.description.ends_with("… · observed 2026-07-01"));
-    assert!(!hit.description.contains("cited"));
-}
-
-#[test]
-fn symbol_hits_take_the_bare_name_and_the_file_location() {
-    let frame = contextgraph_types::ContextFrame {
-        id: "code-graph:sym:src/lib.rs:12:issue_row".into(),
-        kind: contextgraph_types::FrameKind::Symbol,
-        title: "fn issue_row".into(),
-        content: Some("fn issue_row(...) { ... }".into()),
-        uri: Some("file:///repo/src/lib.rs".into()),
-        score: 0.9,
-        token_cost: 10,
-        content_digest: None,
-        representation: contextgraph_types::Representation::Full,
-        content_fidelity: None,
-        canonical_content_hash: None,
-        content_ref: None,
-        transform: None,
-        minimum_content_fidelity: None,
-        inline_content_requirement: None,
-        canonical_token_cost: None,
-        tokenizer_ref: None,
-
-        valid_from: None,
-        valid_to: None,
-        recorded_at: None,
-        provenance: vec![],
-        citation_label: Some("fn issue_row (src/lib.rs:12)".into()),
-        embedding: None,
-        relations: vec![],
-    };
-    let hit = symbol_hit(&frame);
-    assert_eq!(hit.kind, "Symbol");
-    assert_eq!(hit.label, "fn issue_row");
-    assert_eq!(hit.insert, "issue_row", "the bare name is what inserts");
-    assert_eq!(hit.description, "src/lib.rs:12");
-
-    // Without a citation label the frame's uri stands in.
-    let mut bare = frame;
-    bare.citation_label = None;
-    assert_eq!(symbol_hit(&bare).description, "file:///repo/src/lib.rs");
-}
-
-#[test]
-fn merge_assignee_hits_orders_agents_then_local_and_caps() {
-    let person = |l: &str| EntityHit {
-        kind: "Person".into(),
-        label: l.into(),
-        description: String::new(),
-        insert: l.into(),
-    };
-    let agents: Vec<EntityHit> = (0..2).map(|i| person(&format!("a{i}"))).collect();
-    let local: Vec<EntityHit> = (0..3).map(|i| person(&format!("m{i}"))).collect();
-    let merged = merge_assignee_hits(agents, local, 4);
-    let labels: Vec<&str> = merged.iter().map(|h| h.label.as_str()).collect();
-    assert_eq!(
-        labels,
-        vec!["a0", "a1", "m0", "m1"],
-        "agents first, then local — capped"
-    );
-}
-
-#[test]
-fn local_assignee_hits_read_as_empty_on_a_bare_workspace() {
-    // Read-only politeness: no `.stella/` databases → no hits and, above
-    // all, no directories/files created as a side effect.
-    let dir = tempfile::tempdir().unwrap();
-    assert!(local_assignee_hits(dir.path(), "anything").is_empty());
-    assert!(
-        !dir.path().join(".stella").exists(),
-        "a lookup must never create the workspace store"
-    );
-}
 
 /// `requeue_front` front-inserts in push order and mirrors every insert
 /// to the deck as `PromptRequeued`, so the driver's backlog and the
